completed implementation of locationHistory route functions

+++ routes/userRoutes/locationHistory.js
@@ -8,42 +8,56 @@ const LocationHistoryEntry = require("../../models/LocationHistoryEntry");
 
 router.get(
     "/entries",
-    (async(req, res) => {
+    (async (req, res) => {
         // Return list of user's location entries.
         const currentUser = req.session.username;
-        const entries = await User.findOne(
-            {username : currentUser}
-        ).select("locationHistory");
-        //console.log(`entries: ${entries}`)
+        const userObj = await User.findOne(
+            { username: currentUser }
+        );
+
+        let entries = [];
+        for (let i = 0; i < userObj.locationHistory.length; i++) {
+            entries.push(await LocationHistoryEntry.findById(userObj.locationHistory[i]));
+        }
+
         res.status(200).send(entries);
     })
 );
 
 router.post(
     "/addEntry",
-    (async(req, res) => {
+    (async (req, res) => {
         // Add an entry to the user's location history.
         const data = req.body;
         const currentUser = req.session.username;
         const userObj = await User.findOne(
-            {username : currentUser}
+            { username: currentUser }
         );
 
-        try 
-        {
+        try {
+            // Check that the entry doesn't already exist.
+            if (await LocationHistoryEntry.exists({
+                user: userObj._id,
+                businessLocation: data.businessLocation,
+                timeIn: data.timeIn,
+                timeOut: data.timeOut
+            })) {
+                return res.status(400).json({
+                    message: "Entry already exists",
+                });
+            }
+
             const newEntry = await LocationHistoryEntry.create({
                 user: userObj._id,
                 businessLocation: data.businessLocation,
                 timeIn: data.timeIn,
                 timeOut: data.timeOut
             });
-            // await newEntry.save();
             userObj.locationHistory.push(newEntry._id);
             await userObj.save();
-        } 
-        catch (err) 
-        {
-            return res.status(500).json({message: "Unable to create entry", error: err});
+        }
+        catch (err) {
+            return res.status(500).json({ message: "Unable to create entry", error: err });
         }
 
         res.sendStatus(200);
@@ -52,26 +66,60 @@ router.post(
 
 router.post(
     "/editEntry",
-    (async(req, res) => {
+    (async (req, res) => {
         // Edit an entry in the user's location history.
+        const data = req.body;
+        const currentUser = req.session.username;
+        const userObj = await User.findOne(
+            { username: currentUser }
+        );
 
+        try {
+            const entryId = new mongoose.Types.ObjectId(req.body.id);
+            // Check that the location exists.
+            if (!await LocationHistoryEntry.exists({ _id: entryId })) {
+                return res.status(400).json({
+                    message: "Location does not exist",
+                });
+            }
+
+            let entry = await LocationHistoryEntry.findById(entryId);
+            entry.set({
+                user: userObj._id,
+                businessLocation: data.businessLocation,
+                timeIn: data.timeIn,
+                timeOut: data.timeOut
+            });
+            await entry.save();
+
+        } catch (error) {
+            return res.status(500).json({ message: "Not able to update entry.", error: err });
+        }
+
+        res.status(200).json({ message: "Location updated." });
     })
 );
 
 router.delete(
     "/deleteEntry",
-    (async(req, res) => {
+    (async (req, res) => {
         // Delete an entry in the user's location history.
         const currentUser = req.session.username;
         const userObj = await User.findOne(
-            {username : currentUser}
+            { username: currentUser }
         );
 
         try {
             // Try to delete the document.
             const entryId = new mongoose.Types.ObjectId(req.body.id);
+            // Check that the location exists.
+            if (!await LocationHistoryEntry.exists({ _id: entryId })) {
+                return res.status(400).json({
+                    message: "Entry does not exist",
+                });
+            }
 
-            userObj.locationHistory.pull({_id: entryId});
+            userObj.locationHistory.pull({ _id: entryId });
             await userObj.save();
 
             await LocationHistoryEntry.deleteOne({
@@ -80,7 +128,7 @@ router.delete(
         } catch (err) {
             return res.status(500).json({ message: "Not able to delete entry.", error: err });
         }
-    
+
         res.status(200).json({ message: "Entry deleted." });
     })
 );
